Issue #123: Refactored the splitGet function into a http transport component, added tests.
